requirements.txt
# torch == 2.0.1
# torchvision == 0.15.2
# numpy == 1.26.1
opencv-python
gdown
tqdm
psutil

# droid-slam
torch-scatter
open3d
tensorboard
scipy
matplotlib
pyyaml
evo
lietorch
plyfile

# depth-pro
timm
pillow
